+++ src/index.ts
@@ -6,6 +6,7 @@ import { AttributesAPI } from "./api/attributes";
 import { BlogsAPI } from "./api/blogs";
 import { CategoriesAPI } from "./api/categories";
 import { CharitiesAPI } from "./api/charities";
+import { CurrencyAPI } from "./api/currencies";
 import { DonationsAPI } from "./api/donations";
 import { EventsAPI } from "./api/events";
 import { MessagesAPI } from "./api/messages";
@@ -25,6 +26,7 @@ export class MarkkoSDK {
   public blogs: BlogsAPI;
   public categories: CategoriesAPI;
   public charities: CharitiesAPI;
+  public currencies: CurrencyAPI;
   public donations: DonationsAPI;
   public events: EventsAPI;
   public messages: MessagesAPI;
@@ -44,6 +46,7 @@ export class MarkkoSDK {
     this.blogs = new BlogsAPI(config, this.authService);
     this.categories = new CategoriesAPI(config, this.authService);
     this.charities = new CharitiesAPI(config, this.authService);
+    this.currencies = new CurrencyAPI(config, this.authService);
     this.donations = new DonationsAPI(config, this.authService);
     this.events = new EventsAPI(config, this.authService);
     this.messages = new MessagesAPI(config, this.authService);