+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.0-alpha.xx]-2025-02-10
+
+### Added
+
+- Currency API endpoints:
+  - Retrieve available currencies
+
 ## [1.0.0-alpha.35]-2025-02-07
 
 ### Added